Implement support for PRBuilder and Maven too
As of this commit this tool can be used for Maven and the
SparkPullRequestBuilder in addition to for SBT only.

This commit also changes the semantics of the "Count" column
represented on the spreadsheet in the following way: now it is
the number of runs a particular test suite as failed in, as
opposed to the number of tests within that suite over all runs.
This prevents a single build that fails many tests in the same
suite (e.g. HiveCompatibilitySuite) to DoS the script.