Add skeleton for separating data collection from reporting
The previous design highly intertwines data collection (fetching
failed suites) and reporting (posting to spreadsheet). Reporting
results remotely real time is not really a killer feature but
significantly complicates the code and limits its evolution.

The new architecture separates these two aspects. The data
collection continues to go through the JSON/REST API that Jenkins
exposes, but logs failure information to a CSV file locally instead.
The reporting then can do anything with the failure information
(e.g. aggregating statistics about a test suite or across a Jenkins
project) without further complicating the fetching code.

One thing this allows us to do in the future is to track an
individual test suite over the course of a long time. This would
have been virtually impossible to do in the old architecture.

Note that this commit adds only a skeleton because the reporting
side is not implemented yet. That will come in a future commit.