Set up Google spreadsheet reporting framework
This commit sets up the framework for us to report test failure
information through Google spreadsheet. The architecture further
separates computation from reporting. The former is handled in
the parent TestFailureReporter class, and the latter is handled
in the child GoogleSpreadsheetReporter class. The harness provided
by this commit will be extended to actually report comprehensive
test failure information in the next commit.