Implement complete reporting framework
This commit implements all the missing logic needed to report test
failure information to Google spreadsheets. Under the hood it uses
SparkSQL to perform computations on the output file produced in
the fetching phase. The reporting phase is decoupled to provide
flexibility on the kind of computation that can be done on these
test results.

This commit also moves all fetching logic from Python to Scala
to simplify configuration. The next commit will add a README that
makes it even simpler for the user to use this tool.